handle api images and homepage

+++ app/Http/Controllers/Admins/ProductController.php
@@ -5,16 +5,13 @@ namespace App\Http\Controllers\Admins;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admins\ProductRequest;
 use App\Http\Resources\Admins\ProductsResource;
-use App\Models\Product;
-use App\Models\Variation;
 use App\Models\Color;
+use App\Models\Product;
 use App\Models\Size;
 use App\Models\SizeGuide;
-
 use App\Models\Tag;
 use DB;
 use Illuminate\Http\Request;
-use Illuminate\Support\Str;
 use Image;
 
 class ProductController extends Controller
@@ -23,18 +20,21 @@ class ProductController extends Controller
     {
 
         $this->middleware([
-        'admin']);
+            'admin']);
 
     }
 
-
-    
     public function deleteSizeGuides($id)
     {
-      
-        SizeGuide::where('product_id' , $id)->delete();
-    }
 
+        SizeGuide::where('product_id', $id)->delete();
+    }
+    public function deleteProductVariations($id)
+    {
+
+        DB::table('variations')->where('product_id', $id)->delete();
+
+    }
 
     public function getArrayFromString($tagsNamesArray)
     {
@@ -45,20 +45,25 @@ class ProductController extends Controller
         return $arrayNames;
     }
 
-    public function updateData($product, $request)
+    public function updateData($product, $request , $id)
     {
         $product->category_id = $request->category_id;
-        $product->price = $request->discount_price ?  $request->discount_price : $request->selling_price ;
+        $product->price = $request->discount_price ? $request->discount_price : $request->selling_price;
         $product->name = $request->name;
         $product->selling_price = $request->selling_price;
         $product->discount_price = $request->discount_price;
         $product->slug = str_replace(" ", "-", $request->name);
         $product->long_description = $request->long_description;
         $product->short_description = $request->short_description;
-        $product->offer = $product->discount_price ? ( ($product->selling_price - $product->discount_price) * 100 ) / $product->selling_price : null ;
+        $product->offer = $product->discount_price ? (($product->selling_price - $product->discount_price) * 100) / $product->selling_price : null;
         $product->sku = $request->sku;
-        $product->sizes =  $request->sizesNamesArray;
-        $product->colors = $request->colorsNamesArray;
+        if($id == null ) { 
+            $product->sizes = $request->sizesNamesArray;
+            $product->colors = $request->colorsNamesArray;
+            $product->wash_care = "Cold water wash";
+            $product->contents = '60% cottons,40% ligra';
+        }
+     
         $product->wash_care = $request->wash_care;
         $product->contents = $request->contents;
     }
@@ -66,21 +71,20 @@ class ProductController extends Controller
     public function addVariations($request, $product)
     {
 
-
-
         $colorsNamesArray = $this->getArrayFromString($request->colorsNamesArray);
+
         $sizesNamesArray = $this->getArrayFromString($request->sizesNamesArray);
 
-        $colorsIds = Color::whereIn('name' , $colorsNamesArray)->pluck('id')->all();
+        $colorsIds = Color::whereIn('name', $colorsNamesArray)->pluck('id')->all();
 
-        $sizesIds = Size::whereIn('name' , $sizesNamesArray)->pluck('id')->all();
+        $sizesIds = Size::whereIn('name', $sizesNamesArray)->pluck('id')->all();
         foreach ($colorsIds as $colorId) {
             foreach ($sizesIds as $sizeId) {
                 DB::table('variations')->insert([
                     'color_id' => $colorId,
                     'size_id' => $sizeId,
                     'product_id' => $product->id,
-                    'sku' => $product->sku .$colorId.$sizeId,
+                    'sku' => $product->sku . $colorId . $sizeId,
                     'selling_price' => $product->selling_price,
                     'discount_price' => $product->discount_price,
                     'image' => '/images/products/variations/variation.jpg',
@@ -92,22 +96,37 @@ class ProductController extends Controller
 
     }
 
-    public function uploadImage($image, $product)
+    public function uploadImage($product, $id)
     {
-        $imageName = $image->getClientOriginalExtension();
-        $imageName = time() . "." . $imageName;
-        Image::make($image)->fit(600, 800)->save(public_path("/images/products/") . $imageName);
-        if ($product->image && file_exists(public_path() . $product->image) && $product->image != "/images/products/product.jpg") {
-            unlink(substr($product->image, 1));
+        $image = request()->file("image");
+        if ($image) {
+            // delete old image
+            if ($id && file_exists(public_path() . $product->image) && $product->image != "/images/products/product.jpg") {
+                unlink(substr($product->image, 1));
+            }
+            // delete old image
+            $imageName = $image->getClientOriginalExtension();
+            $imageName = time() . "." . $imageName;
+            Image::make($image)->fit(600, 800)->save(public_path("/images/products/") . $imageName, 50);
+            if ($product->image && file_exists(public_path() . $product->image) && $product->image != "/images/products/product.jpg") {
+                unlink(substr($product->image, 1));
+            }
+            $product->image = "/images/products/" . $imageName;
         }
-        $product->image = "/images/products/" . $imageName;
+
+        $product->save();
+        $response = [
+            'product' => $product,
+        ];
+
+        return response($response, 201);
     }
 
-    public function assignTags($tagsNamesArray, $product)
+    public function assignTags($request, $product)
     {
         $originalTagsNames = Tag::pluck('name')->all();
 
-        $requestTagsNamesArray = $this->getArrayFromString($tagsNamesArray);
+        $requestTagsNamesArray = $this->getArrayFromString($request->tagsNamesArray);
 
         foreach ($requestTagsNamesArray as $tagName) {
             if (!in_array($tagName, $originalTagsNames)) {
@@ -125,7 +144,7 @@ class ProductController extends Controller
 
     public function index()
     {
-        $products = collect(ProductsResource::collection(Product::orderBy('created_at', 'DESC')->with([ 'category', 'variations', 'tags', 'colors', 'sizes'])->get()));
+        $products = collect(ProductsResource::collection(Product::orderBy('created_at', 'DESC')->with(['category', 'variations', 'tags', 'colors', 'sizes'])->get()));
         $tags = DB::table('tags')->get();
         $response = [
             'products' => $products,
@@ -138,42 +157,20 @@ class ProductController extends Controller
     public function store(ProductRequest $request)
     {
         $product = new product;
-        $image = request()->file("image");
-        $this->updateData($product, $request);
-        $this->uploadImage($image, $product);
+        $this->updateData($product, $request , null);
+        $this->uploadImage($product, null);
+        $this->assignTags($request, $product);
+        $this->addVariations($request, $product);
 
-        $product->save();
-         $this->assignTags($request->tagsNamesArray, $product);
-         $this->addVariations($request, $product);
-        $response = [
-            'product' => $product,
-        ];
-
-        return response($response, 201);
     }
 
     public function update(ProductRequest $request, $id)
     {
 
         $product = Product::find($id);
-        $this->updateData($product, $request);
-
-        $image = request()->file("image");
-
-        if ($image) {
-            $this->uploadImage($image, $product);
-        }
-
-        $product->save();
-
-        $this->assignTags($request->tagsNamesArray, $product);
-
-        $response = [
-            'product' => $product,
-        ];
-
-        return response($response, 201);
-
+        $this->updateData($product, $request , $id);
+        $this->uploadImage($product, $id);
+        $this->assignTags($request, $product);
     }
 
     public function changeStatusOrFeaturedOrTrend(Request $request, $id)
@@ -210,19 +207,18 @@ class ProductController extends Controller
 
         $product = Product::find($id);
 
-        if ($product->image && file_exists(public_path() . $product->image) && $product->image != "/images/products/product.jpg") {
-            $imageDeleted = unlink(substr($product->image, 1));
+        if (file_exists(public_path() . $product->image) && $product->image != "/images/products/product.jpg") {
+            $imageFileDeleted = unlink(substr($product->image, 1));
 
-            if ($imageDeleted) {
-
+            if ($updateStatus) {
                 $product->delete();
                 $this->deleteProductVariations($id);
-               $this->deleteSizeGuides($id);
+                $this->deleteSizeGuides($id);
             }
 
         } else {
             $product->delete();
-             $this->deleteProductVariations($id);
+            $this->deleteProductVariations($id);
             $this->deleteSizeGuides($id);
         }
 
@@ -234,10 +230,4 @@ class ProductController extends Controller
 
     }
 
-    public function deleteProductVariations($id)
-    {
-
-        DB::table('variations')->where('product_id' , $id)->delete();
-
-    }
 }
